pyproject.toml
[project]
name = "tesla-fsd-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cantools>=40.2.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "python-can-csscan-mf4>=2.4.0",
    "python-can[mf4]>=4.5.0",
]
